+++ source/Loom.Tests/EventSourcing/DelegatingEventReader.cs
@@ -6,16 +6,16 @@
 
     internal class DelegatingEventReader : IEventReader
     {
-        private readonly Func<Guid, int, Task<IEnumerable<object>>> _function;
+        private readonly Func<Guid, long, Task<IEnumerable<object>>> _function;
 
         public DelegatingEventReader(
-            Func<Guid, int, Task<IEnumerable<object>>> function)
+            Func<Guid, long, Task<IEnumerable<object>>> function)
         {
             _function = function;
         }
 
         public Task<IEnumerable<object>> QueryEventPayloads(
-            Guid streamId, int afterVersion)
+            Guid streamId, long afterVersion)
         {
             return _function.Invoke(streamId, afterVersion);
         }
